convert action app to a plain python package, move action django stuff to project app
