+++ baleen/project/urls.py
@@ -1,5 +1,6 @@
 from django.conf.urls import patterns, url
 
+# Projects 
 urlpatterns = patterns('',
     url(r'^$', 'baleen.project.views.index', name='project_index'),
     url(r'^add$', 'baleen.project.views.add', name='add_project'),
@@ -7,6 +8,7 @@ urlpatterns = patterns('',
     url(r'^(?P<project_id>\d+)/deploy$', 'baleen.project.views.manual_deploy', name='manual_deploy'),
 )
 
+# Jobs
 urlpatterns += patterns('',
     url(r'^(?P<project_id>\d+)/job/(?P<job_id>\d+)$',
         'baleen.job.views.view_job', name='view_job'),
@@ -19,3 +21,13 @@ urlpatterns += patterns('',
         'baleen.job.views.view_specific_action_html_coverage',
         name='view_specific_action_html_coverage'),
 )
+
+# Actions
+urlpatterns = patterns('',
+    url(r'^(?P<project_id>\d+)/action$', 'baleen.action.views.add_action', name='add_action'),
+    url(r'^(?P<project_id>\d+)/action/(?P<action_id>\d+)$',
+        'baleen.action.views.edit_action', name='edit_action'),
+    url(r'^(?P<project_id>\d+)/action-order$',
+        'baleen.action.views.set_action_order', name='set_action_order'),
+)
+
