+++ baleen/settings.py
@@ -194,6 +194,12 @@ GEARMAN_JOB_LABEL = 'baleen_job'
 
 GITHUB_HOOK_URL = SITE_URL + '/hub'
 
+ACTION_MODULES = {
+    'project': "baleen.action.project",
+    'docker': "baleen.action.docker",
+    'ssh': "baleen.action.ssh",
+}
+
 # A sample logging configuration. The only tangible logging
 # performed by this configuration is to send an email to
 # the site admins on every HTTP 500 error when DEBUG=False.
